requirements.txt
requests
python-dotenv
fastapi
uvicorn
apscheduler
shutil
